[fixed] warnings solved
git-svn-id: http://svn2.assembla.com/svn/swgemu/engine3public@156 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOEngine/include/engine/service/DatagramServiceClientProxy.h
@@ -22,7 +22,7 @@ namespace engine {
 		}
 
 		bool send(Packet* pack) {
-			if (packetLossChance != 0 && System::random(100) < packetLossChance)
+			if (packetLossChance != 0 && System::random(100) < (uint32) packetLossChance)
 				return false;
 			
 			socket->sendTo(pack, &addr);
@@ -31,7 +31,7 @@ namespace engine {
 		}
 
 		bool read(Packet* pack) {
-			if (packetLossChance != 0 && System::random(100) < packetLossChance)
+			if (packetLossChance != 0 && System::random(100) < (uint32) packetLossChance)
 				return false;
 
 			SocketAddress addr;
